+++ src/debounce/mod.rs
@@ -150,7 +150,7 @@ impl Debounce {
 
     pub fn configure(&mut self, config: Config, tx: Sender<Result<bool>>) {
         tx.send(match config {
-            Config::OngoingWrites(c) => self.timer.set_ongoing_writes(c),
+            Config::OngoingEvents(c) => self.timer.set_ongoing(c),
             _ => Ok(false),
         })
         .expect("configuration channel disconnected");